Mobile enhancement
Add mobile enhancement on tap to update the info window, using the click event. The update required the Leaflet library at version 0.7.3 as to not conflict with the d3 data. Also, reformat the CSS style and update the html document with the end tags for the html's body and end html tags.